feat(reward): implement agentfs-sdk integration
Integrates agentfs-sdk (v0.6.4) into the reward system to fetch real-world
tool statistics from AgentFS SQLite databases.

- Replaced AgentFsProvider stubs with real SDK calls in memory-core.
- Optimized tool stat fetching using batch queries (tc.stats()).
- Corrected SignalMerger confidence logic to return 0.0 for empty signals.
- Updated MCP server tools in memory-mcp to utilize the integrated SDK.
- Fixed multiple compilation errors in tests, improved HashSet performance,
  and removed unused imports to satisfy Clippy.
- Verified merger logic with unit tests; core build is verified via targeted
  checks despite local environment limitations.

Fixes the TODO in memory-core/src/reward/external/agentfs.rs.
