+++ memory-core/src/memory/retrieval/scoring.rs
@@ -199,8 +199,6 @@ mod tests {
             metadata: std::collections::HashMap::new(),
             tags: Vec::new(),
             checkpoints: Vec::new(),
-            version: 1,
-            parent_id: None,
         })
     }
