+++ memory-core/src/patterns/clustering/tests.rs
@@ -153,8 +153,6 @@ mod tests {
             salient_features: None,
             tags: vec![],
             checkpoints: vec![],
-            version: 1,
-            parent_id: None,
             metadata: std::collections::HashMap::new(),
         };
