+++ memory-core/src/retrieval/cascade/mod.rs
@@ -9,6 +9,8 @@
 //! The cascade eliminates 50-70% of embedding API calls by satisfying
 //! queries from CPU-local tiers before falling back to the API.
 
+use anyhow::Result;
+
 mod concept_graph;
 pub use concept_graph::ConceptGraph;
 
@@ -196,7 +198,7 @@ impl CascadeRetriever {
     /// 4. API fallback (requires external embedding call)
     ///
     /// Without `csm`, returns empty results (placeholder behavior).
-    pub fn retrieve(&self, query: &str) -> CascadeResult {
+    pub fn retrieve(&self, query: &str) -> Result<CascadeResult> {
         #[cfg(feature = "csm")]
         {
             self.retrieve_with_csm(query)
@@ -207,18 +209,18 @@ impl CascadeRetriever {
             // Placeholder implementation - returns empty results
             // query is intentionally unused in placeholder mode
             let _ = query;
-            CascadeResult {
+            Ok(CascadeResult {
                 episode_ids: Vec::new(),
                 scores: Vec::new(),
                 contributing_tiers: Vec::new(),
                 api_calls: 0,
-            }
+            })
         }
     }
 
     /// Full cascade implementation using CSM components.
     #[cfg(feature = "csm")]
-    fn retrieve_with_csm(&self, query: &str) -> CascadeResult {
+    fn retrieve_with_csm(&self, query: &str) -> Result<CascadeResult> {
         use super::{compute_weights, merge_results};
 
         // Tier 1: BM25 keyword search
@@ -226,12 +228,12 @@ impl CascadeRetriever {
 
         // Check if BM25 produced sufficient results
         if bm25_results.sufficient {
-            return CascadeResult {
+            return Ok(CascadeResult {
                 episode_ids: bm25_results.ids(),
                 scores: bm25_results.scores(),
                 contributing_tiers: vec!["bm25".to_string()],
                 api_calls: 0,
-            };
+            });
         }
 
         // Tier 2: HDC similarity search
@@ -245,20 +247,20 @@ impl CascadeRetriever {
 
             // Check if merged results are sufficient
             if merged.len() >= self.config.min_results {
-                return CascadeResult {
+                return Ok(CascadeResult {
                     episode_ids: merged.iter().map(|(id, _)| id.clone()).collect(),
                     scores: merged.iter().map(|(_, s)| *s).collect(),
                     contributing_tiers: vec!["bm25".to_string(), "hdc".to_string()],
                     api_calls: 0,
-                };
+                });
             }
         } else if hdc_results.sufficient {
-            return CascadeResult {
+            return Ok(CascadeResult {
                 episode_ids: hdc_results.ids(),
                 scores: hdc_results.scores(),
                 contributing_tiers: vec!["hdc".to_string()],
                 api_calls: 0,
-            };
+            });
         }
 
         // Tier 3: ConceptGraph expansion (optional)
@@ -266,12 +268,12 @@ impl CascadeRetriever {
             let concept_results = self.retrieve_concept_graph(query);
 
             if concept_results.sufficient {
-                return CascadeResult {
+                return Ok(CascadeResult {
                     episode_ids: concept_results.ids(),
                     scores: concept_results.scores(),
                     contributing_tiers: vec!["concept_graph".to_string()],
                     api_calls: 0,
-                };
+                });
             }
         }
 
@@ -286,7 +288,7 @@ impl CascadeRetriever {
             bm25_results.results.clone()
         };
 
-        CascadeResult {
+        Ok(CascadeResult {
             episode_ids: best_results.iter().map(|(id, _)| id.clone()).collect(),
             scores: best_results.iter().map(|(_, s)| *s).collect(),
             contributing_tiers: if best_results.is_empty() {
@@ -304,7 +306,7 @@ impl CascadeRetriever {
                 tiers
             },
             api_calls: 1, // Indicates API call would be needed
-        }
+        })
     }
 
     /// BM25 keyword search (Tier 1).
