+++ memory-core/src/retrieval/cascade/tests.rs
@@ -57,7 +57,7 @@ fn test_placeholder_retrieve_without_csm() {
     {
         let mut retriever = CascadeRetriever::default_config();
         retriever.add_episode("ep-1", "Test episode");
-        let result = retriever.retrieve("test query");
+        let result = retriever.retrieve("test query").unwrap();
         // Without CSM feature, returns empty results
         assert!(result.episode_ids.is_empty());
         assert!(result.scores.is_empty());
@@ -120,7 +120,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "error handling patterns in async code");
 
         // Query with exact keyword match should return 0 API calls
-        let result = retriever.retrieve("authentication JWT token");
+        let result = retriever.retrieve("authentication JWT token").unwrap();
 
         // Should find the matching episode
         assert!(!result.episode_ids.is_empty());
@@ -145,7 +145,9 @@ mod csm_tests {
         retriever.add_episode("ep-5", "fix memory leak in cache implementation");
 
         // Semantic-like query (similar words but not exact match)
-        let result = retriever.retrieve("user authentication and login security");
+        let result = retriever
+            .retrieve("user authentication and login security")
+            .unwrap();
 
         // Should find related episodes
         assert!(!result.episode_ids.is_empty());
@@ -163,7 +165,7 @@ mod csm_tests {
         let retriever = CascadeRetriever::default_config();
 
         // Empty index should indicate API call needed
-        let result = retriever.retrieve("any query");
+        let result = retriever.retrieve("any query").unwrap();
 
         assert!(result.episode_ids.is_empty());
         // Should indicate API fallback needed
@@ -180,11 +182,11 @@ mod csm_tests {
         retriever.add_episode("ep-3", "unique_keyword_gamma optimization");
 
         // Query that matches exactly should hit BM25
-        let result = retriever.retrieve("unique_keyword_alpha");
+        let result = retriever.retrieve("unique_keyword_alpha").unwrap();
         assert!(result.contributing_tiers.contains(&"bm25".to_string()));
 
         // Query with no exact match but similar content should use HDC
-        let result = retriever.retrieve("implement alpha feature");
+        let result = retriever.retrieve("implement alpha feature").unwrap();
         // Either BM25 (if partial match) or HDC (if semantic similarity)
         assert!(!result.episode_ids.is_empty() || result.api_calls > 0);
     }
@@ -210,7 +212,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "async task spawning performance tips");
 
         // Query that matches multiple aspects
-        let result = retriever.retrieve("Rust async programming");
+        let result = retriever.retrieve("Rust async programming").unwrap();
 
         // Should have results from merging BM25 and HDC
         assert!(!result.episode_ids.is_empty());
@@ -236,7 +238,7 @@ mod csm_tests {
         retriever.add_episode("ep-1", "authentication implementation");
         retriever.add_episode("ep-2", "database connection setup");
 
-        let result = retriever.retrieve("authentication");
+        let result = retriever.retrieve("authentication").unwrap();
 
         // Without merge, should only use single tier
         assert!(result.contributing_tiers.len() <= 1);
@@ -272,7 +274,7 @@ mod csm_tests {
         retriever.add_episode("ep-2", "database pool connection");
         retriever.add_episode("ep-3", "rate limiting API");
 
-        let result = retriever.retrieve("authentication");
+        let result = retriever.retrieve("authentication").unwrap();
 
         // All scores should be in 0.0-1.0 range
         for score in &result.scores {
@@ -293,7 +295,7 @@ mod csm_tests {
             retriever.add_episode(&format!("ep-{i}"), &format!("episode {i} content"));
         }
 
-        let result = retriever.retrieve("episode");
+        let result = retriever.retrieve("episode").unwrap();
 
         // Should not exceed top_k
         assert!(result.episode_ids.len() <= 3);
@@ -348,7 +350,7 @@ mod csm_tests {
         retriever.add_episode("ep-5", "refactor error handling patterns");
 
         // Query with abbreviated terms that need expansion
-        let result = retriever.retrieve("fix auth bug");
+        let result = retriever.retrieve("fix auth bug").unwrap();
 
         // Should find results via concept graph expansion ("auth" → authentication domain)
         assert!(!result.episode_ids.is_empty());
