+++ memory-core/src/types/emitter.rs
@@ -175,21 +175,6 @@ impl MemoryEventMapping {
                     "extraction_timestamp": timestamp,
                 }),
             ),
-            MemoryEvent::ConceptDriftDetected {
-                parent_id,
-                version_count,
-                changepoint_count,
-                timestamp,
-            } => (
-                "com.do-memory.concept.drift_detected".to_string(),
-                parent_id.clone(),
-                serde_json::json!({
-                    "parent_id": parent_id,
-                    "version_count": version_count,
-                    "changepoint_count": changepoint_count,
-                    "drift_timestamp": timestamp,
-                }),
-            ),
         }
     }
 }
@@ -415,29 +400,4 @@ mod tests {
         assert_eq!(deserialized.specversion, "1.0");
         assert_eq!(deserialized.event_type, "com.test");
     }
-}
-
-#[test]
-fn test_concept_drift_event_mapping() {
-    use super::MemoryEventMapping;
-    use crate::types::event::MemoryEvent;
-    use uuid::Uuid;
-
-    let parent_id = Uuid::new_v4().to_string();
-    let timestamp = 1_714_500_400;
-    let mem_event = MemoryEvent::ConceptDriftDetected {
-        parent_id: parent_id.clone(),
-        version_count: 5,
-        changepoint_count: 2,
-        timestamp,
-    };
-
-    let (event_type, subject, data) = MemoryEventMapping::map_event(&mem_event);
-
-    assert_eq!(event_type, "com.do-memory.concept.drift_detected");
-    assert_eq!(subject, parent_id);
-    assert_eq!(data["parent_id"], parent_id);
-    assert_eq!(data["version_count"], 5);
-    assert_eq!(data["changepoint_count"], 2);
-    assert_eq!(data["drift_timestamp"], timestamp);
 }
