+++ memory-core/src/types/event.rs
@@ -47,17 +47,6 @@ pub enum MemoryEvent {
         /// Unix timestamp in seconds
         timestamp: u64,
     },
-    /// Concept drift was detected in a versioned episode chain.
-    ConceptDriftDetected {
-        /// Parent episode ID (the "concept" identifier)
-        parent_id: String,
-        /// Number of versions analyzed
-        version_count: u32,
-        /// Number of changepoints detected
-        changepoint_count: u32,
-        /// Unix timestamp in seconds
-        timestamp: u64,
-    },
 }
 
 impl MemoryEvent {
@@ -68,8 +57,7 @@ impl MemoryEvent {
             Self::EpisodeCreated { timestamp, .. }
             | Self::EpisodeCompleted { timestamp, .. }
             | Self::EpisodeGarbageCollected { timestamp, .. }
-            | Self::PatternExtracted { timestamp, .. }
-            | Self::ConceptDriftDetected { timestamp, .. } => *timestamp,
+            | Self::PatternExtracted { timestamp, .. } => *timestamp,
         }
     }
 
@@ -81,7 +69,6 @@ impl MemoryEvent {
             | Self::EpisodeCompleted { id, .. }
             | Self::EpisodeGarbageCollected { id, .. } => id,
             Self::PatternExtracted { id, .. } => id,
-            Self::ConceptDriftDetected { parent_id, .. } => parent_id,
         }
     }
 }
