+++ memory-mcp/src/mcp/tools/mod.rs
@@ -4,7 +4,6 @@
 
 pub mod advanced_pattern_analysis;
 pub mod checkpoint;
-pub mod concept_drift;
 pub mod embeddings;
 pub mod episode_relationships;
 pub mod episode_tags;
