+++ memory-storage-redb/tests/cache_integration_test.rs
@@ -40,8 +40,6 @@ fn create_test_episode(id: Uuid) -> Episode {
         metadata: HashMap::new(),
         tags: vec![],
         checkpoints: vec![],
-        version: 1,
-        parent_id: None,
     }
 }
