+++ memory-storage-turso/src/tests.rs
@@ -561,8 +561,6 @@ mod compression_tests {
             applied_patterns: vec![],
             salient_features: None,
             tags: vec![],
-            version: 1,
-            parent_id: None,
             checkpoints: vec![],
             start_time: chrono::Utc::now(),
             end_time: None,
@@ -605,8 +603,6 @@ mod compression_tests {
             patterns: vec![],
             heuristics: vec![],
             applied_patterns: vec![],
-            version: 1,
-            parent_id: None,
             salient_features: None,
             tags: vec![],
             checkpoints: vec![],
