+++ memory-storage-turso/tests/phase1_optimization_test.rs
@@ -49,8 +49,6 @@ fn create_test_episode(id: Uuid, domain: &str) -> Episode {
         metadata,
         tags: vec![],
         checkpoints: vec![],
-        version: 1,
-        parent_id: None,
     }
 }
